requirements.txt
pandas
numpy
matplotlib
scikit-learn
xgboost
mlforecast
streamlit
requests
websocket-client
python-dotenv
